requirements.txt
asgiref==3.2.10
Django==3.0.5
django-cors-headers==3.4.0
django-mysql==3.7.1
djangorestframework==3.11.0
djongo==1.3.3
mysqlclient==2.0.0
PyJWT==1.7.1
pymongo==3.10.1
pytz==2020.1
sqlparse==0.3.1
